build: strict options and build directory
The build system puts compiled class files now into the `bld` directory.
Why `bld, you ask? For the same reason sources are kept in `src`. This
is one of those old conventions worth discussing. Is there a benefit of
`src` over `sources`?

Furthermore, we add a bunch of strict options regarding warning, docs
and linting, that `javac` checks for us.
It is intentionally too strict now and will loosen up over time.

The build is hard-coded to be a "debug" build for now.