adds minimal entry point
This commit gets the "hello world" of java modules going.
The build is now setup in a way, that the "main" module is the root node
of the compilation graph.
That means, as long as other sources and referenced modules live in
src/, no further modification of the build is necessary.

Building independent modules however, will require future modifications.

The name of the main module is likely to change in the future, but for
kick-off it suffices.

This commit also contains the result of experimentation to get the wall
time of the build down.
It seems clear, that the stated goal of "1 sec build" will be hard to
maintain.
Currently, this "hello world"-type of project needs ~1.3s on my x230 and
~3s on my e6400.
Setting the heap space and some other options for the runtime shaves off
a couple of milliseconds. See the `x` file.

Pre-compiling the build with `javac` cuts the time roughly in half.

I will further investigate, if code modification enhance the time, but
it seems unlikely, given that this is pretty minimal already.